requirements.txt
# 설치 라이브러리
streamlit
google-generativai
python-dotenv
pandas
#sqlite3 = 기본 라이브러리

# 설치 명령어
# pip install streamlit google-generativeai python-dotenv pandas
